pyproject.toml
[tool.poetry]
name = "besta-tools"
version = "0.1.0"
description = "Tools for working with Besta data formats"
authors = ["dogtopus <dogtopus@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pefile = "^2021.9.3"
pyelftools = "^0.27"
click = "^8.1.7"
construct-typing = "^0.6.2"

[tool.poetry.dev-dependencies]
rope = "^0.20.1"
pylint = "^2.11.1"
mypy = "^0.910"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
elf2bestape = 'besta_tools.elf2bestape:main'
titlemgr = 'besta_tools.titlemgr:main'
